I've tried to enable Swagger (Swashbuckle) in .Net Core for OData. It seems I can't have both unfortunately. Swashbuckle has a .Net Core version that does not support OData, only standard WebAPI controllers. And they have an OData version that does not support .Net Core (only .Net Framework 4.6.1+).
Numerous people have asked for support for both, and there are a number of partial workarounds, but nothing offering a proper solution and the 'owner' of Swashbuckle has not indicated whether he plan to develop this.

So for now we wait, eventually something will surface one way or the other.

https://github.com/rbeauchamp/Swashbuckle.OData/issues/154

#65